+++ noodles-cram/examples/cram_read_header.rs
@@ -12,9 +12,7 @@ fn main() -> io::Result<()> {
     let src = env::args().nth(1).expect("missing src");
 
     let mut reader = cram::reader::Builder::default().build_from_path(src)?;
-    reader.read_file_definition()?;
-
-    let header = reader.read_file_header()?;
+    let header = reader.read_header()?;
     print!("{header}");
 
     Ok(())