+++ noodles-cram/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+  * cram/reader: Add `Reader::read_header`.
+
+    This is a convenience method that reads the file definition and header and
+    returns a parsed SAM header.
+
 ## 0.33.0 - 2023-06-15
 
 ### Changed